index page 2

+++ database/migrations/2020_02_29_132004_create_abonners_table.php
@@ -15,7 +15,7 @@ class CreateAbonnersTable extends Migration
     {
         Schema::create('abonners', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('num')->unsigned()->unique();
+            $table->bigInteger('num')->unsigned()->unique();
             $table->string('nom',20);
             $table->string('prenom',20);
             $table->string('password',200);
